+++ src/main/scala/com/github/karlhigley/spark/neighbors/lsh/ScalarRandomProjectionFunction.scala
@@ -38,7 +38,7 @@ private[neighbors] class ScalarRandomProjectionFunction(
   /**
    * Build a hash table entry for the supplied vector
    */
-  def hashTableEntry(id: Int, table: Int, v: SparseVector): IntHashTableEntry = {
+  def hashTableEntry(id: Long, table: Int, v: SparseVector): IntHashTableEntry = {
     IntHashTableEntry(id, table, signature(v), v)
   }
 }
