+++ src/main/scala/com/github/karlhigley/spark/neighbors/lsh/SignRandomProjectionFunction.scala
@@ -37,7 +37,7 @@ private[neighbors] class SignRandomProjectionFunction(
   /**
    * Build a hash table entry for the supplied vector
    */
-  def hashTableEntry(id: Int, table: Int, v: SparseVector): BitHashTableEntry = {
+  def hashTableEntry(id: Long, table: Int, v: SparseVector): BitHashTableEntry = {
     BitHashTableEntry(id, table, signature(v), v)
   }
 }
